fix: next/headers Client Component import エラーの修正
### 問題
- auth-context.tsx (Client Component) から helpers.ts (Server Component) をimport
- helpers.ts が next/headers をimportすることでビルドエラー発生

### 解決策
- client-helpers.ts を新規作成してクライアント専用関数を分離
- types.ts を作成して型定義を共通化
- auth-context.tsx のimportを client-helpers.ts に変更

### 変更ファイル
- 新規: src/lib/auth/client-helpers.ts (クライアント専用ヘルパー)
- 新規: src/lib/auth/types.ts (共通型定義)
- 修正: src/lib/auth/helpers.ts (サーバー専用に整理)
- 修正: src/lib/auth/auth-context.tsx (import修正)
- 修正: src/lib/auth/__tests__/helpers.test.ts (import修正)

### テスト結果
- 全テスト通過: 81 passed, 1 skipped ✅
- ビルド成功: 12ページ生成完了 ✅
