+++ src/lib/auth/helpers.ts
@@ -1,13 +1,6 @@
 import { cookies } from 'next/headers'
 import { NextRequest } from 'next/server'
-
-// MVP架構用的用户类型
-interface MVPUser {
-  id: string
-  email: string
-  role: string
-  created_at: string
-}
+import { MVPUser } from './types'
 
 // サーバーコンポーネント用認証チェック
 export async function getServerSession(): Promise<MVPUser | null> {
@@ -88,58 +81,8 @@ export async function authenticateAdmin(req: NextRequest): Promise<MVPUser> {
   return session
 }
 
-// クライアントサイド用セッション設定
-export function setClientSession(user: MVPUser) {
-  if (typeof window !== 'undefined') {
-    // Cookie設定 (httpOnly=falseでクライアントからアクセス可能)
-    document.cookie = `auth-session=authenticated; path=/; max-age=${60 * 60 * 24 * 7}` // 7日間
-    document.cookie = `user-data=${encodeURIComponent(JSON.stringify(user))}; path=/; max-age=${60 * 60 * 24 * 7}`
-    document.cookie = `user-role=${user.role}; path=/; max-age=${60 * 60 * 24 * 7}`
-  }
-}
-
-// クライアントサイド用セッション削除
-export function clearClientSession() {
-  if (typeof window !== 'undefined') {
-    document.cookie =
-      'auth-session=; path=/; expires=Thu, 01 Jan 1970 00:00:01 GMT'
-    document.cookie =
-      'user-data=; path=/; expires=Thu, 01 Jan 1970 00:00:01 GMT'
-    document.cookie =
-      'user-role=; path=/; expires=Thu, 01 Jan 1970 00:00:01 GMT'
-  }
-}
-
-// セッション状態チェック (クライアントサイド)
-export function getClientSession(): MVPUser | null {
-  if (typeof window === 'undefined') {
-    return null
-  }
-
-  try {
-    const cookies = document.cookie.split(';')
-    const authCookie = cookies.find((c) => c.trim().startsWith('auth-session='))
-    const userDataCookie = cookies.find((c) =>
-      c.trim().startsWith('user-data=')
-    )
-
-    if (!authCookie || !userDataCookie) {
-      return null
-    }
-
-    const authValue = authCookie.split('=')[1]
-    if (authValue !== 'authenticated') {
-      return null
-    }
-
-    const userDataValue = userDataCookie.split('=')[1]
-    const userData = JSON.parse(decodeURIComponent(userDataValue)) as MVPUser
-    return userData
-  } catch (error) {
-    console.error('Client session error:', error)
-    return null
-  }
-}
+// Note: Client-side session functions moved to client-helpers.ts
+// to avoid next/headers import issues in Client Components
 
 // プロファイル取得 (MVP: 基本情報のみ)
 export async function getUserProfile(userId: string): Promise<MVPUser | null> {
@@ -168,14 +111,7 @@ export async function refreshSession(): Promise<MVPUser | null> {
   }
 }
 
-// 認証状態を同期 (クライアントサイドとサーバーサイド)
-export function syncAuthState(user: MVPUser | null) {
-  if (user) {
-    setClientSession(user)
-  } else {
-    clearClientSession()
-  }
-}
+// Note: syncAuthState removed - use client-helpers directly for client-side operations
 
 // ユーザーロール判定ヘルパー
 export function isAdmin(user: MVPUser | null): boolean {
